feat: move base image check to public pkg
