+++ pkg/container/allowed_base_image_check_test.go
@@ -27,7 +27,7 @@ import (
 )
 
 func TestShouldPassIfNoDockerfilePresent(t *testing.T) {
-	result := NewAllowedBaseImage().Test()
+	result := NewAllowedBaseImage("./").Test()
 
 	if result == nil || result.Passed == false {
 		t.Errorf("Allowed base image check should pass, if there is no Dockerfile found")
@@ -39,7 +39,7 @@ func TestShouldFailIfDockerfileWithUnapprovedBaseImagePresent(t *testing.T) {
 	_ = file.writeTo("./")
 	defer os.Remove(file.filename)
 
-	result := NewAllowedBaseImage().Test()
+	result := NewAllowedBaseImage("./").Test()
 
 	if result.Passed {
 		t.Errorf("Allowed based image check should fail, if Dockerfile with unapproved base image found")
@@ -51,7 +51,7 @@ func TestShouldPassIfTemurinIsUsedAsBaseImage(t *testing.T) {
 	_ = file.writeTo("./")
 	defer os.Remove(file.filename)
 
-	if !(NewAllowedBaseImage().Test().Passed) {
+	if !(NewAllowedBaseImage("./").Test().Passed) {
 		t.Errorf("eclipse/temurin should be recognized as approved base image")
 	}
 }
@@ -64,7 +64,7 @@ func TestShouldNotFailIfOnlyBuildLayerDeviatesFromTemurin(t *testing.T) {
 	_ = file.writeTo("./")
 	defer os.Remove(file.filename)
 
-	if !(NewAllowedBaseImage().Test().Passed) {
+	if !(NewAllowedBaseImage("./").Test().Passed) {
 		t.Errorf("Unapproved images in build layers should not let the check fail")
 	}
 }
@@ -75,7 +75,7 @@ func TestShouldFailForUnallowedBaseImageInSubdirectory(t *testing.T) {
 	_ = dockerfile.writeTo(subdirectory)
 	defer os.RemoveAll(subdirectory)
 
-	result := NewAllowedBaseImage().Test()
+	result := NewAllowedBaseImage("./").Test()
 
 	if result.Passed {
 		t.Errorf("Unapproved base images should also be detected in sub directories")
@@ -93,7 +93,7 @@ func TestShouldFailIfAtLeastOneDockerfileWithUnallowedBaseImageIsFound(t *testin
 	defer os.RemoveAll(firstSubdir)
 	defer os.RemoveAll(secondSubDir)
 
-	result := NewAllowedBaseImage().Test()
+	result := NewAllowedBaseImage("./").Test()
 
 	if result.Passed {
 		t.Errorf("Base image check should fail, if at least one Dockerfile with unallowed base image is found!")
@@ -112,7 +112,7 @@ func TestShouldAllowBaseImagesFromWhitelist(t *testing.T) {
 	defer os.RemoveAll(firstSubdir)
 	defer os.RemoveAll(secondSubDir)
 
-	result := NewAllowedBaseImage().Test()
+	result := NewAllowedBaseImage("./").Test()
 
 	if !result.Passed {
 		t.Errorf("Should allow base images from whitelist")
@@ -127,7 +127,7 @@ func TestShouldIncludeAllUnallowedBaseImagesInErrorDescription(t *testing.T) {
 	defer os.RemoveAll(firstSubdir)
 	defer os.RemoveAll(secondSubDir)
 
-	result := NewAllowedBaseImage().Test()
+	result := NewAllowedBaseImage("./").Test()
 
 	if !strings.Contains(result.ErrorDescription, "badBaseImage") || !strings.Contains(result.ErrorDescription, "another/unallowed") {
 		t.Errorf("Error message should contain all denied base images")
